Fix to avoid polluting $error on missing modules.
When a completer has a script block to determine
the commands to register for, the script block
might not find any commands, in which case we'd
try to register a completer with $null for the
command name and that's not allowed.

Fix is to skip registering anything in this case.